requirements.txt
fastapi
uvicorn
gunicorn
pydantic
joblib
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
azure-identity
azure-keyvault-secrets